+++ CHANGELOG.md
@@ -31,6 +31,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 #### Added
 
 - [Hero] New information element below the hero for photographer credits
+- [StatusLabel] Added SCSS support and scss files are exported too. See documentation about usage.
 
 #### Changed
 
@@ -92,7 +93,7 @@ Changes that are not related to specific components
 
 #### Added
 
-- [Component] What is added?
+- [StatusLabel] Documentation about SCSS usage.
 
 #### Changed