+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.10.1] - Oct, 09, 2024
+
+### React
+
+#### Fixed
+
+- [Logo] Fixed data-testid attribute not being set with native props
+- [Login] Fixed GraphQLModule exports
+
 ## [3.10.0] - Sep, 04, 2024
 
 ### React